fix(cli): installation detection and homebrew support
Resolves #581
